Custom Field Checker V1 Upload
This Update Set contains the Custom Field Checker functionality that allows you to quickly check the amount of and usage of fields on your tables. Whilst primarily used for custom fields, this can also be used to view any fields on a table that meet the search query you define.